requirements.txt
# Core dependencies
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0
PyYAML>=6.0,<7.0.0

# Apache Spark (optional - for distributed execution)
pyspark>=3.4.0,<4.0.0          
delta-spark>=2.4.0,<3.0.0      

# Kafka (optional - for real streaming)
# confluent-kafka>=2.2.0,<3.0.0
kafka-python>=2.0.0,<3.0.0

# Monitoring and logging
mlflow>=2.8.0,<3.0.0

# Testing
pytest>=7.4.0,<9.0.0
pytest-cov>=4.1.0,<6.0.0
pytest-mock>=3.11.0,<4.0.0

# Code quality
flake8>=6.1.0,<8.0.0
black>=23.9.0,<25.0.0
isort>=5.12.0,<6.0.0
mypy>=1.5.0,<2.0.0

# Documentation
# sphinx>=7.0.0,<8.0.0

# HTTP client for Jira integration
requests>=2.31.0,<3.0.0

psycopg2-binary>=2.9.0,<3.0.0